etl/requirements.txt
elasticsearch>=8.11.0
psycopg2-binary>=2.9.9
urllib3>=2.0.0
